+++ Box.V2/Managers/BoxUsersManager.cs
@@ -217,13 +217,15 @@ namespace Box.V2.Managers
         }
 
         /// <summary>
-        /// Retrieves information about a user in the enterprise. Requires enterprise administration authorization.
+        /// Retrieves information about a user in the enterprise with the specified fields. Requires enterprise administration authorization.
         /// </summary>
         /// <param name="userId">The user identifier.</param>
-        /// <returns>Returns the complete user object.</returns>
-        public async Task<BoxUser> GetUserInformationAsync(string userId)
+        /// <param name="fields">Attribute(s) to include in the response.</param>
+        /// <returns>Returns the default representation of the user object.</returns>
+        public async Task<BoxUser> GetUserInformationAsync(string userId, IEnumerable<string> fields = null)
         {
-            BoxRequest request = new BoxRequest(_config.UserEndpointUri, userId);
+            BoxRequest request = new BoxRequest(_config.UserEndpointUri, userId)
+            .Param(ParamFields, fields);
 
             IBoxResponse<BoxUser> response = await ToResponseAsync<BoxUser>(request).ConfigureAwait(false);