+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## 3.21.0 [2019-12-05]
+- Added `fields` parameter to `UsersManager.GetUserInformationAsync()`
+- Added `ExternalAppUserId` property to `BoxUser` model
+- Added the ability to set the `TrackingCodes` property when updating or creating a user (thanks @Cpcrook!)
+
+## 3.20.0 [2019-09-19]
+- Added missing fields for File Version object.
+
+## 3.19.0 [2019-08-29]
+- Added `FILE_VERSION_RESTORE` constant to Admin Event.
+- Added action_by field to Enterprise Event.
+- Audited missing fields on BoxFile and BoxFolder objects.
+- Better error handling and messaging for errors pertaining to OAuth2 error responses.
+
+## 3.18.0 [2019-06-20]
+
+- Added `sort` and `direction` parameters to `FoldersManager.GetFolderItemsAsync()` to enable
+  sorting the folder items returned
+- Added a new `SearchManager.QueryAsync()` method with correct types for file size filter parameters
+- Deprecated the `SearchManager.SearchAsync()` method, which is superseded by `SearchManager.QueryAsync()`
+- Added support for setting the `IsExternalCollabRestricted` parameter when creating and updating Users
+- Added a `WebProxy` property to `BoxConfig` instances, which can be used to manually set the network proxy
+  used by the SDK
+
 ## 3.17.0 [2019-05-09]
 
 - Fixed the encoding of dates in the query parameters for Events and Search endpoints